Db migration

+++ backend/atividade05/questao01/src/db/migrations/20210703180806-create-produto.js
@@ -15,7 +15,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       preco: {
-        type: Sequelize.REAL
+        type: Sequelize.INTEGER
       },
       createdAt: {
         allowNull: false,
